requirements.txt
python-telegram-bot==20.7
Pillow
qrcode
pyzbar
APScheduler
python-dotenv